using waveguide obj for ring
This would allow the user to also define sidewall angles.
It should be noted bezier curve for cirlces are approximate. I used a loose approximation that did not leave gaps between the quadrants (it was necessary to copy and rotate quarter circles, as the waveguide obj doesn't allow overlap for some reason).
If more resolution and smoothing is needed, the bezier poles need to be defined better.
Alternatively, the user can instead use the Ring obj, which would be an exact ring but would not allow sidewalls.